requirements.txt
streamlit>=1.34
httpx>=0.27
eth-abi>=5.1
eth-utils>=4.0
eth-hash[pycryptodome]>=0.7
pandas>=2.2
altair>=5.3
structlog>=24.4
pytest>=8.3
pytest-asyncio>=0.23
ruff>=0.6
mypy>=1.11
python-dotenv>=1.0

